refactor: entity 필드를 원시타입으로 변경

+++ module-market/src/main/java/com/market/board/controller/MarketBoardContentController.java
@@ -28,7 +28,7 @@ public class MarketBoardContentController {
     }
 
     @GetMapping
-    public MarketBoardContentDto.Response getMarketBoardcontent(int mbcSeq) {
+    public MarketBoardContentDto.Response getMarketBoardcontent(Long mbcSeq) {
         return MarketBoardContentDto.from(marketBoardContentService.find(mbcSeq));
     }
 }
