+++ module-market/src/main/java/com/market/board/repository/MarketBoardRepository.java
@@ -3,5 +3,5 @@ package com.market.board.repository;
 import com.market.board.entity.MarketBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarketBoardRepository extends JpaRepository<MarketBoard, Integer> {
+public interface MarketBoardRepository extends JpaRepository<MarketBoard, Long> {
 }
