+++ module-market/src/main/java/com/market/board/repository/MarketBoardStateRepository.java
@@ -3,5 +3,5 @@ package com.market.board.repository;
 import com.market.board.entity.MarketBoardState;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarketBoardStateRepository extends JpaRepository<MarketBoardState, Integer> {
+public interface MarketBoardStateRepository extends JpaRepository<MarketBoardState, Long> {
 }
